Add Filename field to doc.Source
The doc package assumed that the source file URL looks like a file name
ending with ".go".  This is wrong in the case of Google Code Hosting
where subrepo source URLs end with a query string.

Add Filename field to doc.Source so that the doc builder can examine the
filename.