Ensure that meta requests include a non empty path.
This is the correct thing to do and fixes a problem with a poorly
configured server. The server in question returns differnt documents for
https://domain, https://domain/ and http://domain/. The server also has
an invalid certificate. The "go get" command skates through the mess
because it errors out on the https requests. The production urlfetch
service does not error out on these requests.